Move example auth route into a subfolder
